+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Enable the use of markdown extensions in package READMEs (#236)
 * Autofocus the search field on the home page (#235)
 * Support release changelogs (#241)
+* Packages are no longer kept as their own dependent (#242)
 
 ## v1.0.4 -- 2022-10-02